# swagger.yaml
{
  "openapi": "3.0.2",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/users/get_users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Read Users",
        "operationId": "read_users_users_get_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Users Users Get Users Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/add_user": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add User",
        "operationId": "add_user_users_add_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Add User Users Add User Post",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/is_newbie/{tg_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Is Newbie",
        "operationId": "is_newbie_users_is_newbie__tg_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Tg Id",
              "type": "string"
            },
            "name": "tg_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Is Newbie Users Is Newbie  Tg Id  Get",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/get_tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Read Tasks",
        "operationId": "read_tasks_tasks_get_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Tasks Tasks Get Tasks Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/get_task/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Read Task",
        "operationId": "read_task_tasks_get_task__task_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            },
            "name": "task_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/create_task": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task Handler",
        "operationId": "create_task_handler_tasks_create_task_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/edit_task/{task_id}": {
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "Update Task",
        "operationId": "update_task_tasks_edit_task__task_id__put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            },
            "name": "task_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/": {
      "get": {
        "tags": [
          "assignments"
        ],
        "summary": "Read Assignments",
        "operationId": "read_assignments_assignments__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Assignments Assignments  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Assignment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/assignments/submit_assignment": {
      "post": {
        "tags": [
          "assignments"
        ],
        "summary": "Submit Assignment",
        "operationId": "submit_assignment_assignments_submit_assignment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submit Assignment Assignments Submit Assignment Post",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/{assignment_id}": {
      "get": {
        "tags": [
          "assignments"
        ],
        "summary": "Read Assignment",
        "operationId": "read_assignment_assignments__assignment_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "integer"
            },
            "name": "assignment_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/results/{tg_id}": {
      "get": {
        "tags": [
          "assignments"
        ],
        "summary": "Get Best Results",
        "operationId": "get_best_results_assignments_results__tg_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Tg Id",
              "type": "string"
            },
            "name": "tg_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Best Results Assignments Results  Tg Id  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Assignment"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assignments/assignments/{task_id}/{tg_id}": {
      "get": {
        "tags": [
          "assignments"
        ],
        "summary": "Get User Assignments By Task",
        "operationId": "get_user_assignments_by_task_assignments_assignments__task_id___tg_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            },
            "name": "task_id",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Tg Id",
              "type": "string"
            },
            "name": "tg_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get User Assignments By Task Assignments Assignments  Task Id   Tg Id  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Assignment"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Assignment": {
        "title": "Assignment",
        "required": [
          "task_id",
          "tg_id",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          },
          "tg_id": {
            "title": "Tg Id",
            "type": "string"
          },
          "mark": {
            "title": "Mark",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "error_message": {
            "title": "Error Message",
            "type": "string"
          },
          "assignment_file": {
            "title": "Assignment File",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AssignmentBase": {
        "title": "AssignmentBase",
        "required": [
          "assignment_id",
          "task_id"
        ],
        "type": "object",
        "properties": {
          "assignment_id": {
            "title": "Assignment Id",
            "type": "integer"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          },
          "mark": {
            "title": "Mark",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "error_message": {
            "title": "Error Message",
            "type": "string"
          }
        }
      },
      "AssignmentCreate": {
        "title": "AssignmentCreate",
        "required": [
          "task_id",
          "tg_id"
        ],
        "type": "object",
        "properties": {
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          },
          "tg_id": {
            "title": "Tg Id",
            "type": "string"
          },
          "mark": {
            "title": "Mark",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "error_message": {
            "title": "Error Message",
            "type": "string"
          },
          "assignment_file": {
            "title": "Assignment File",
            "type": "string"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Task": {
        "title": "Task",
        "required": [
          "task_name",
          "task_status",
          "task_description",
          "task_type",
          "task_data",
          "requirements",
          "task_id"
        ],
        "type": "object",
        "properties": {
          "task_name": {
            "title": "Task Name",
            "type": "string"
          },
          "task_status": {
            "title": "Task Status",
            "type": "string"
          },
          "task_description": {
            "title": "Task Description",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          },
          "task_data": {
            "title": "Task Data",
            "type": "string"
          },
          "requirements": {
            "title": "Requirements",
            "type": "string"
          },
          "task_data_admin": {
            "title": "Task Data Admin",
            "type": "string"
          },
          "check_solution_file": {
            "title": "Check Solution File",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          }
        }
      },
      "TaskCreate": {
        "title": "TaskCreate",
        "required": [
          "task_name",
          "task_status",
          "task_description",
          "task_type",
          "task_data",
          "requirements"
        ],
        "type": "object",
        "properties": {
          "task_name": {
            "title": "Task Name",
            "type": "string"
          },
          "task_status": {
            "title": "Task Status",
            "type": "string"
          },
          "task_description": {
            "title": "Task Description",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          },
          "task_data": {
            "title": "Task Data",
            "type": "string"
          },
          "requirements": {
            "title": "Requirements",
            "type": "string"
          },
          "task_data_admin": {
            "title": "Task Data Admin",
            "type": "string"
          },
          "check_solution_file": {
            "title": "Check Solution File",
            "type": "string"
          }
        }
      },
      "TaskUpdate": {
        "title": "TaskUpdate",
        "type": "object",
        "properties": {
          "task_name": {
            "title": "Task Name",
            "type": "string"
          },
          "task_status": {
            "title": "Task Status",
            "type": "string"
          },
          "task_description": {
            "title": "Task Description",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          },
          "task_data": {
            "title": "Task Data",
            "type": "string"
          },
          "requirements": {
            "title": "Requirements",
            "type": "string"
          },
          "task_data_admin": {
            "title": "Task Data Admin",
            "type": "string"
          },
          "check_solution_file": {
            "title": "Check Solution File",
            "type": "string"
          }
        }
      },
      "User": {
        "title": "User",
        "required": [
          "tg_id",
          "auth_timestamp"
        ],
        "type": "object",
        "properties": {
          "tg_id": {
            "title": "Tg Id",
            "type": "string",
            "description": "Идентификатор пользователя в Telegram (tg_id)"
          },
          "username": {
            "title": "Username",
            "type": "string",
            "description": "Уникальное имя пользователя в системе"
          },
          "auth_timestamp": {
            "title": "Auth Timestamp",
            "type": "string",
            "description": "Время, когда пользователь начал взаимодействовать с ботом",
            "format": "date-time"
          },
          "assignments": {
            "title": "Assignments",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssignmentBase"
            },
            "description": "Список заданий, связанных с пользователем",
            "default": []
          }
        }
      },
      "UserCreate": {
        "title": "UserCreate",
        "required": [
          "tg_id"
        ],
        "type": "object",
        "properties": {
          "tg_id": {
            "title": "Tg Id",
            "type": "string",
            "description": "Идентификатор пользователя в Telegram (tg_id)"
          },
          "username": {
            "title": "Username",
            "type": "string",
            "description": "Уникальное имя пользователя в системе"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}